+++ dflib-echarts/src/main/java/org/dflib/echarts/saver/HtmlChartModel.java
@@ -3,5 +3,5 @@ package org.dflib.echarts.saver;
 /**
  * @since 2.0.0
  */
-public record HtmlChartModel(String chartDiv, String chartScript) {
+public record HtmlChartModel(String id, String chartDiv, String chartScript) {
 }